remove the view modes
there was the concept of show effects (plugins) mode or edit mode.
however, I think it does not serve a purpose at all and one does not see
any plugins that are in fact active on the signal being processed in
edit mode.
Better to give up some screenspace then not realizing a plugin is
messing up your work ...